pyproject.toml
[project]
name = "metatomic-lj-test"
description = "Package for testing metatomic model integration"
dynamic = ["version"]
dependencies = [
    "torch >=2.1",
    "metatomic-torch >=0.1.6"
]

[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "cmake",
    "setuptools",
    "setuptools-scm",
    "torch >=2.1"
]

[tool.setuptools_scm]
